backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
SQLAlchemy==2.0.36
psycopg2-binary==2.9.10
pydantic-settings==2.7.1
python-multipart==0.0.20
pandas==2.2.3
numpy==2.2.1
scikit-learn==1.6.0
python-dateutil==2.9.0.post0
